// src/app/model/constancia.model.ts
export class Constancia {
    id: string;
    cveTipo: string;
    tipo: string;
    folio: string;
    institucion: string;
    ejercicioFiscal: number;
    cuenta: string;
    rfc: string;
    cod_cliente: string;
    nombre: string;
    participacion_cliente: string;
    estatus: string;
    disabled: boolean;
}